+++ test/cypress/integration/baseR4.current-version.e2e.cy.ts
@@ -1,150 +1,59 @@
-import {
-  MatStepHarness,
-  MatStepperHarness,
-  MatStepperNextHarness
-} from '@angular/material/stepper/testing';
-import { getHarness } from '@jscutlery/cypress-harness';
-import 'cypress-file-upload';
-
 describe('Research Data Finder (baseR4)', () => {
-  // Page objects & harnesses
-  // See https://material.angular.io/cdk/test-harnesses/overview for details
-  let stepper: MatStepperHarness;
-  let stepsArray: Array<MatStepHarness>;
-  let settingsStep: MatStepHarness;
-  let selectAnActionStep: MatStepHarness;
-  let selectRecordsStep: MatStepHarness;
-  let viewCohortStep: MatStepHarness;
-  let pullDataStep: MatStepHarness;
 
-  before(() => {
-    cy.visit('/?server=https://lforms-fhir.nlm.nih.gov/baseR4&prev-version=disable')
-      .get('app-fhir-server-select.loading')
-      .should('exist')
-      .get('app-fhir-server-select.loading', {timeout: 30000})
+  it( 'should show spinner during server initialization', () => {
+    cy.visit('/?server=https://lforms-fhir.nlm.nih.gov/baseR4&prev-version=disable');
+    cy.get('app-fhir-server-select.loading')
+      .should('exist');
+    cy.get('app-fhir-server-select.loading', {timeout: 30000})
       .should('not.exist');
-
-    // Initialize common page objects (harnesses)
-    getHarness(MatStepperHarness)
-      .then((result) => {
-        stepper = result;
-        return stepper.getSteps();
-      })
-      .then((stepsArr) => {
-        stepsArray = stepsArr;
-        [
-          settingsStep,
-          selectAnActionStep,
-          selectRecordsStep,
-          viewCohortStep,
-          pullDataStep
-        ] = stepsArray;
-      });
   });
 
-  // Current step next button harness
-  let nextPageBtn: MatStepperNextHarness;
-
-  beforeEach((done) => {
-    stepper
-      .getSteps({selected: true})
-      .then(([currentStep]) =>
-        currentStep
-          ? currentStep.getHarness(MatStepperNextHarness).catch(() => null)
-          : null
-      )
-      .then((btn) => {
-        nextPageBtn = btn;
-        done();
-      });
+  it('should display welcome message for the first step', () => {
+    cy.contains('app-stepper > p:first-child', 'This is a query tool')
+      .should('be.visible');
   });
 
-  it('should display welcome message', () => {
-    cy.get('app-stepper > p:first-child').should('be.visible');
+  it('should display 2 steps at the beginning', () => {
+    cy.checkStepCount(2);
   });
 
-  it('should display 2 steps', () => {
-    expect(stepsArray.length).to.equal(2);
+  it('should select the "Settings" step by default', () => {
+    cy.isStepSelected('Settings');
   });
 
-  it('should select the Settings step by default', (done) => {
-    settingsStep.isSelected().then((isSelected) => {
-      expect(isSelected).to.equal(true);
-      done();
-    });
+  it('should allow to proceed to the "Select An Action" step', () => {
+    cy.clickButton( 'Select an action');
+    cy.isStepSelected( 'Select an action');
   });
 
-  it('should allow to proceed to the Select An Action step', (done) => {
-    nextPageBtn
-      .click()
-      .then(() => selectAnActionStep.isSelected())
-      .then((isSelected) => {
-        expect(isSelected).to.equal(true);
-        done();
-      });
+  it('should not display welcome message for the next steps', () => {
+    cy.contains('app-stepper > p:first-child', 'This is a query tool')
+      .should('not.exist');
   });
 
-  it('should display all steps after selecting cart approach', (done) => {
-    cy.contains(
-      'Create a cohort of patients by browsing and selecting records'
-    ).click();
-    cy.then(() => getHarness(MatStepperHarness))
-      .then((result: MatStepperHarness) => {
-        stepper = result;
-        return stepper.getSteps();
-      })
-      .then((stepsArr) => {
-        stepsArray = stepsArr;
-        [
-          settingsStep,
-          selectAnActionStep,
-          selectRecordsStep,
-          viewCohortStep,
-          pullDataStep
-        ] = stepsArray;
-        expect(stepsArray.length).to.equal(5);
-        done();
-      });
+  it('should display all steps after selecting cart approach', () => {
+    cy.clickLabel('Create a cohort of patients by browsing and selecting records');
+    cy.checkStepCount(5);
   });
 
-  it('should allow to proceed to the Select Records step', (done) => {
-    cy.contains('Select records')
-      .click()
-      .then(() => selectRecordsStep.isSelected())
-      .then((isSelected) => {
-        expect(isSelected).to.equal(true);
-        done();
-      });
+  it('should allow to proceed to the "Select Records" step', () => {
+    cy.clickButton('Next');
+    cy.isStepSelected('Select records');
   });
 
-  it('should not allow skipping the View cohort (search for patients) step', (done) => {
-    viewCohortStep
-      .select()
-      .then(() => selectRecordsStep.isSelected())
-      .then((isSelected) => {
-        expect(isSelected).to.equal(true);
-        done();
-      });
+  it('should not allow skipping patient search before the "View Cohort" step', () => {
+    cy.selectStep('View cohort');
+    cy.isStepSelected('Select records');
   });
 
-  it('should allow to proceed to the View cohort step', (done) => {
-    nextPageBtn
-      .click()
-      .then(() => viewCohortStep.isSelected())
-      .then((isSelected) => {
-        expect(isSelected).to.equal(true);
-        done();
-      });
+  it('should allow to proceed to the "View cohort" step', () => {
+    cy.clickButton('Search for Patients');
+    cy.isStepSelected('View cohort');
   });
 
-  it('should allow to proceed to the Pull Data for cohort step', (done) => {
-    nextPageBtn
-      .click()
-      .then(() => pullDataStep.isSelected())
-      .then((isSelected) => {
-        expect(isSelected).to.equal(true);
-        done();
-      });
+  it('should allow to proceed to the "Pull data for the cohort" step', () => {
+    cy.clickButton('Pull data for the cohort');
+    cy.isStepSelected('Pull data for the cohort');
   });
 
 });