+++ test/cypress/support/commands.ts
@@ -41,3 +41,87 @@
 //
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
+
+declare namespace Cypress {
+  interface Chainable<Subject = any> {
+    /**
+     * Custom command to initialize the application at the specified URL.
+     * Also checks if spinner is displayed during initialization.
+     * @param url - application URL
+     * @example cy.initApp('/?server=https://dbgap-api.ncbi.nlm.nih.gov/fhir/x1&prev-version=disable')
+     */
+    initApp(url: string): Chainable<Subject>;
+    /**
+     * Custom command to click on a visible button with the specified text content.
+     * @param text - text or regular expression
+     * @example cy.clickButton('text')
+     */
+    clickButton(text: string | RegExp): Chainable<Subject>;
+    /**
+     * Custom command to click on a visible label with the specified text content.
+     * @param text - text or regular expression
+     * @example cy.clickLabel('text')
+     */
+    clickLabel(text: string | RegExp): Chainable<Subject>;
+    /**
+     * Custom command to click on a step header with the specified name.
+     * @param name - full step name
+     * @example cy.selectStep('text')
+     */
+    selectStep(name: string): Chainable<Subject>;
+    /**
+     * Custom command to check if a step with the specified name is selected.
+     * @param name - full step name
+     * @example cy.isStepSelected('text')
+     */
+    isStepSelected(name: string): Chainable<Subject>;
+    /**
+     * Custom command to check the number of visible steps.
+     * @param num - number of visible steps
+     * @example cy.checkStepCount(2)
+     */
+    checkStepCount(num: number): Chainable<Subject>;
+    /**
+     * Custom command to type text into an input field.
+     * @param selector - selector for the input field
+     * @param text - some text
+     * @example cy.typeTextToInput('#inputId', 'text')
+     */
+    typeTextToInput(selector: string, text: string): Chainable<Subject>;
+  }
+}
+
+Cypress.Commands.add('initApp', (url) => {
+  cy.visit(url);
+  cy.get('app-fhir-server-select.loading')
+    .should('exist');
+  return cy.get('app-fhir-server-select.loading', {timeout: 30000})
+    .should('not.exist');
+});
+
+Cypress.Commands.add('clickButton', (containedText) => {
+  return cy.contains('button:visible', containedText).click({timeout: 10000});
+});
+
+Cypress.Commands.add('clickLabel', (containedText) => {
+  return cy.contains('label', containedText).click({timeout: 10000});
+});
+
+Cypress.Commands.add('selectStep', (name) => {
+  return cy.get(`.mat-step-header[ng-reflect-label="${name}"]`).click();
+});
+
+Cypress.Commands.add('isStepSelected', (name) => {
+  return cy.get(
+    `.mat-step-header[ng-reflect-label="${name}"][ng-reflect-selected="true"]`
+  ).should('exist');
+});
+
+Cypress.Commands.add('checkStepCount', (num) => {
+  return cy.get('mat-step-header').should('have.length', num);
+});
+
+Cypress.Commands.add('typeTextToInput', (selector, text) => {
+  cy.get(selector).type(text, {force: true});
+  return cy.get(selector).blur({force: true});
+});