+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [7.0.4] - 2024-12-05
+### Fixed
+- Improved stability of e2e tests.
+- Fixed escaping of FHIR search parameter strings: added backslash escaping.
+
 ## [7.0.3] - 2024-11-06
 ### Fixed
 - Checking for an issue with the "not" modifier during initialization might not